gitignore: "vscode" -> ".vscode"
I was accidentally excluding my vscode settings folder, and it took me
a lot of time to discover it was because I typed it wrong on my global
.gitignore file :v

Also improved bash prompt (it shows current git branch instead of
username if inside a git repo folder).

And I also cleared the TODO.tq file.